requirements.txt
Django==1.4
South==0.7.4
django-compressor==1.1.2
django-registration==0.8
django-tastypie==0.9.11
pytz==2012h
django-ignoretests
django_crontab
requests
django-redis
mysql-python
